% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_depth_timeseries.R
\name{animate_depth_timeseries}
\alias{animate_depth_timeseries}
\title{Animate a depth dependent timeseries from a Dlf file}
\usage{
animate_depth_timeseries(dlf, var_name)
}
\arguments{
\item{dlf}{An S4 object of class Dlf}

\item{var_name}{Name of variable in columns}
}
\value{
A plotly figure
}
\description{
Animate a depth dependent timeseries from a Dlf file
}
\examples{
data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, "daily/DailyP/DailyP-Daily-WaterFlux.dlf")
dlf <- read_dlf(path)
animate_depth_timeseries(dlf, "q")
}
