% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlf.R
\name{$,Dlf-method}
\alias{$,Dlf-method}
\title{$ indexing on the data part of the dlf}
\usage{
\S4method{$}{Dlf}(x, name)
}
\arguments{
\item{x}{The dlf object}

\item{name}{Name of column in data part}
}
\description{
$ indexing on the data part of the dlf
}
