% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisyrVis-package.R
\docType{package}
\name{daisyrVis-package}
\alias{daisyrVis}
\alias{daisyrVis-package}
\title{daisyrVis: Visualization of Daisy Log Files}
\description{
Functions for visualizing log files from Daisy simulations.
}
\author{
\strong{Maintainer}: Silas Ørting \email{silas@di.ku.dk} (\href{https://orcid.org/0000-0002-3081-1547}{ORCID})

}
\keyword{internal}
