% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_wide_to_long.R
\name{depth_wide_to_long}
\alias{depth_wide_to_long}
\title{Transform wide format time series to a long format time series
It is assumed that dlf@data only contains \code{time_name} columns and depth
columns with the format
<var_name>_\if{html}{\out{<depth-below-surface>}}
e.g. "q_100"}
\usage{
depth_wide_to_long(dlf, var_name, time_name = "time", depth_name = "z")
}
\arguments{
\item{dlf}{An S4 object of class Dlf}

\item{var_name}{Name of variable in columns}

\item{time_name}{Name of time column}

\item{depth_name}{Name to use for new depth column}
}
\value{
An S4 object of class Dlf

The data of rhe returned Dlf object contains one row for each time/depth
combination.
}
\description{
Transform wide format time series to a long format time series
It is assumed that dlf@data only contains \code{time_name} columns and depth
columns with the format
<var_name>_\if{html}{\out{<depth-below-surface>}}
e.g. "q_100"
}
\examples{
data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, "daily/DailyP/DailyP-Daily-WaterFlux.dlf")
dlf <- read_dlf(path)
dlf@data[1,]
dlf <- daisy_time_to_timestamp(dlf)
dlf <- depth_wide_to_long(dlf, "q")
head(dlf@data)
}
