% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlf_to_csv.R
\name{dlf_to_csv}
\alias{dlf_to_csv}
\title{Write a dlf object as a csv file}
\usage{
dlf_to_csv(dlf, path, sep = ",", include_dlf_header = FALSE)
}
\arguments{
\item{dlf}{Dlf object to write}

\item{path}{Output path to write to}

\item{sep}{Separator to use in csv file}

\item{include_dlf_header}{If TRUE write the Dlf to the begining of the csv
file}
}
\description{
Write a dlf object as a csv file
}
