% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlf_dir.R
\name{drop_dir_from_names}
\alias{drop_dir_from_names}
\title{Remove directory part from dlf names}
\usage{
drop_dir_from_names(dlfs)
}
\arguments{
\item{dlfs}{List of Dlf objects}
}
\value{
List of S4 objects of class Dlf with dir part stripped from names
}
\description{
Remove directory part from dlf names
}
