% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance.R
\name{mass_balance_summary}
\alias{mass_balance_summary}
\title{Mass balance summary for a set of specified variables}
\usage{
mass_balance_summary(dlfs, input, output, content)
}
\arguments{
\item{dlfs}{Either a list of Dlf or a single Dlf}

\item{input}{Name(s) of variable(s) containing mass input}

\item{output}{Name(s) of variable(s) containing mass output}

\item{content}{Name(s) of variable(s) containing mass content}
}
\value{
Either a single list or a list of lists, with the nested lists having
four elements: Inputs, Outputs, InitialContent, FinalContent, and
Balance.
}
\description{
Mass balance summary for a set of specified variables
}
\examples{
data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, "hourly/P2D-Daily-Soil_Chemical_110cm.dlf")
dlf <- read_dlf(path)
input <- c("In_Matrix", "In_Biopores", "External", "Transform", "Tillage")
output <- c("Decompose", "Leak_Matrix", "Leak_Biopores", "Drain_Soil",
            "Drain_Biopores", "Uptake")
content <- c("Content", "Biopores")
mass_balance_summary(dlf, input, output, content)
}
