% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{merge_dlfs}
\alias{merge_dlfs}
\title{Merge a list of Dlfs}
\usage{
merge_dlfs(dlfs, dlf_name_column = "name")
}
\arguments{
\item{dlfs}{List of Dlfs}

\item{dlf_name_column}{Name of column to store name of dlf each row came from}
}
\value{
A Dlf object with the data of the dlfs merged into a single dlf.
}
\description{
Merge a list of Dlfs
}
\examples{
data_dir <- system.file("extdata", package="daisyrVis")
dlfs <- read_dlf_dir(file.path(data_dir, "annual"))
dlfs <- dlfs[startsWith(names(dlfs), "Annual-FN")]
dlfs <- dir_names_to_columns(dlfs)
dlf <- merge_dlfs(dlfs)
head(dlf@data)
}
