% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual.R
\name{plot_annual}
\alias{plot_annual}
\title{Convenience function for plotting annualy logged variables in a bar plot}
\usage{
plot_annual(dlfs, variables, title_suffix = "")
}
\arguments{
\item{dlfs}{Either a list of Dlf or a single Dlf. If a list with of
components, the names are used for the legend.}

\item{variables}{Either a list of variables to plot, or a single variable}

\item{title_suffix}{A string that is appended to the title of all subplots}
}
\value{
ggplot2
}
\description{
Convenience function for plotting annualy logged variables in a bar plot
}
\examples{
# Load dlf data with read_dlf
data_dir <- file.path(system.file("extdata", package="daisyrVis"),
                      "annual/Annual-FN")
dlfs <- read_dlf_dir(data_dir)
# Be careful with variable names. R does not like - so it is replaced with _
vars <- c("Matrix_Leaching", "Crop_Uptake", "Soil_Drain", "Surface_Loss")
plot_annual(dlfs, vars, " - Annual Field Nitrogen")
}
