% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dlf_timeseries.R
\name{plot_dlf_timeseries}
\alias{plot_dlf_timeseries}
\title{Plot one or more timepoints from one or more dlfs containing depth dependent
timeseries}
\usage{
plot_dlf_timeseries(
  dlfs,
  time_points,
  x_var,
  y_var = "z",
  time_var = "time",
  x_label = "",
  y_label = "",
  legend_name = "dlf",
  title = ""
)
}
\arguments{
\item{dlfs}{An S4 object of class Dlf or a list of Dlf objects}

\item{time_points}{List or vector of time points to plot. All dlfs must have
data for all time points}

\item{x_var}{Name of variable on x axis}

\item{y_var}{Name of variable on y axis}

\item{time_var}{Name of time variable}

\item{x_label}{Label for x axis (unit will be appended)}

\item{y_label}{Label for y axis (unit will be appended)}

\item{legend_name}{Name to use for legend}

\item{title}{Plot title (time point will be appended)}
}
\value{
ggplot2 object
}
\description{
Plot one or more timepoints from one or more dlfs containing depth dependent
timeseries
}
