% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_many.R
\name{plot_many}
\alias{plot_many}
\title{Plot one or more variables from one or more Dlfs with a user supplied geom.
Each variable is plotted in a separate subplot.}
\usage{
plot_many(dlfs, x_var, y_vars, geom, title_suffix = "")
}
\arguments{
\item{dlfs}{Either a list of Dlf or a single Dlf. If a list with of
components, the names are  used for the legend.}

\item{x_var}{Name of variable for x axis}

\item{y_vars}{Either a list of variables to plot, or a single variable}

\item{geom}{A single parameter function calling a ggplot2 geom_* function.
It is passed the plot object as the only parameter.}

\item{title_suffix}{A string this is appended to the title of all subplots}
}
\value{
ggplot2
}
\description{
Plot one or more variables from one or more Dlfs with a user supplied geom.
Each variable is plotted in a separate subplot.
}
\examples{
data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, "annual/Annual-FN/HourlyP-Annual-FN-2-2b.dlf")
dlf <- read_dlf(path)
geom <- function(gg) { gg + ggplot2::geom_col(position="dodge") }
plot_many(dlf, "year", "Crop", geom)
}
