% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_and_lines.R
\name{points_and_lines}
\alias{points_and_lines}
\title{Plot one or more variables from one or more dlfs using points and lines}
\usage{
points_and_lines(dlfs, x_var, y_vars, title_suffix = "", plot_line = FALSE)
}
\arguments{
\item{dlfs}{Either a list of Dlf or a single Dlf. If a list, the names are
used for the legend.}

\item{x_var}{Name of variable for x axis}

\item{y_vars}{Either a list of variables to plot, or a single variable}

\item{title_suffix}{A string that is appended to the title of all subplots}

\item{plot_line}{If TRUE plot both points and line otherwise only plot points}
}
\value{
ggplot2
}
\description{
Plot one or more variables from one or more dlfs using points and lines
}
\examples{
## Load dlf data with read_dlf
data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir,
                  "annual/Annual-Tracer/HourlyP-Annual-Tracer-2-2b.dlf")
dlf <- read_dlf(path)
dlf <- daisy_time_to_timestamp(dlf, "date")
points_and_lines(dlf, "date", "Soil", "Soil tracer", FALSE)
points_and_lines(dlf, "date", "Soil", "Soil tracer", TRUE)
}
