% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlf_dir.R
\name{read_daisy_spawn_output}
\alias{read_daisy_spawn_output}
\title{Walk a directory tree generated by daisy spawn, read all dlf files in it,
and merge similar log files into one dlf
#'}
\usage{
read_daisy_spawn_output(directory, pattern = ".*\\\\.dlf")
}
\arguments{
\item{directory}{Path to directory}

\item{pattern}{Regex pattern of files to include}
}
\value{
A list of S4 objects of class Dlf. Each dlf is named with the
log type
}
\description{
Walk a directory tree generated by daisy spawn, read all dlf files in it,
and merge similar log files into one dlf
#'
}
