% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlf.R
\name{read_dlf}
\alias{read_dlf}
\title{Read a dlf file (Daisy log file) or a directory of dlf files}
\usage{
read_dlf(
  path,
  mode = "auto",
  pattern = ".*\\\\.dlf",
  col_name = "sim",
  convert_time = TRUE
)
}
\arguments{
\item{path}{Path to dlf file or directory containing dlf files.}

\item{mode}{One of 'auto', 'file', 'spawn', 'dir. Default is 'auto'.
If 'auto' try to guess the mode.
If 'file' assume path is a single dlf file.
If 'spawn' assume path is a directory containing log directories generated by
the daisy spawn program.
If 'dir' assume path is a directory tree containing dlf files}

\item{pattern}{Regex pattern of files to include. Only relevant for modes
'auto', 'spawn', and 'dir'}

\item{col_name}{Name of column to store directory name in. Only relevant for
modes 'auto' and 'spawn'.}

\item{convert_time}{If TRUE convert disy time to a timestamp stored under
column 'time' in the data slot of each dlf}
}
\value{
For mode 'file': An S4 object of class Dlf with three slots: header,
units, data.
header is list with everything in the dlf file before the data. key: value
pairs from the header are stored as named components of the list.
units is a data.frame containing the units of the values in data.
data is a data.frame containing the logged values

For mode 'spawn': A named list of Dlf objects, where names are names of log
types
For mode 'dir': A named list of Dlf objects, where names are paths to each
file.
}
\description{
Read a dlf file (Daisy log file) or a directory of dlf files
}
\examples{
data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, "annual/Annual-FN/HourlyP-Annual-FN-2-2b.dlf")
dlf <- read_dlf(path)
slotNames(dlf)
names(dlf@header)
dlf@units$Crop
dlf$Crop # equivalent to dlf@data$Crop
dlf[["Crop"]]
}
\seealso{
read_dlf_file, read_dlf_dir, read_dlf_spawn
}
