% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlf_dir.R
\name{read_dlf_spawn}
\alias{read_dlf_spawn}
\title{Walk a directory tree generated by daisy spawn, read all dlf files in it,
and merge similar log files into one dlf.}
\usage{
read_dlf_spawn(directory, pattern = ".*\\\\.dlf", col_name = "sim")
}
\arguments{
\item{directory}{Path to directory}

\item{pattern}{Regex pattern of files to include}

\item{col_name}{Name of column to store directory name in}
}
\value{
A list of S4 objects of class Dlf. Each dlf is named with the
log type and the directory is stored in the column \code{col_name}
}
\description{
Walk a directory tree generated by daisy spawn, read all dlf files in it,
and merge similar log files into one dlf.
}
\examples{
data_dir <- file.path(system.file("extdata", package="daisyrVis",
                                  "daisy-spawn-like"))
print(list.files(data_dir, recursive=TRUE))
dlfs <- read_dlf_spawn(data_dir)
dlf_names <- names(dlfs)
print(dlf_names)

print(unique(dlfs[[dlf_names[1]]]$sim))
}
