% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlf_dir.R
\name{strip_common_prefix_from_names}
\alias{strip_common_prefix_from_names}
\title{Strip common prefix from dlf names}
\usage{
strip_common_prefix_from_names(dlfs)
}
\arguments{
\item{dlfs}{List of Dlf objects}
}
\value{
List of Dlf objects with common prefix stripped from names
}
\description{
Strip common prefix from dlf names
}
