% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlf.R
\name{[[,Dlf-method}
\alias{[[,Dlf-method}
\title{[[ indexing on the data part of the dlf}
\usage{
\S4method{[[}{Dlf}(x, i, j)
}
\arguments{
\item{x}{The dlf object}

\item{i}{Name or index of row OR if j is missing, name or index of column}

\item{j}{Name or index of column}
}
\description{
[[ indexing on the data part of the dlf
}
