% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlf.R
\docType{class}
\name{Dlf-class}
\alias{Dlf-class}
\alias{Dlf}
\title{Class encapsulating Daisy log file (dlf) data}
\description{
Class encapsulating Daisy log file (dlf) data
}
\section{Slots}{

\describe{
\item{\code{header}}{A list with named components containing meta information}

\item{\code{units}}{A data.frame with the units of the values in body}

\item{\code{body}}{A data.frame with the values from the dlf file}
}}

