% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual.R
\name{plot_annual}
\alias{plot_annual}
\title{Make a bar plot of one or more variables from one or more dlfs}
\usage{
plot_annual(dlfs, variables, title_suffix = "")
}
\arguments{
\item{dlfs}{Either a list of Dlf or a single Dlf. If a list with of
components, the names are used for the legend.}

\item{variables}{Either a list of variables to plot, or a single variable}

\item{title_suffix}{A string that is appended to the title of all subplots}
}
\value{
ggplot2
}
\description{
Make a bar plot of one or more variables from one or more dlfs
}
\examples{
# Load dlf data with read_dlf
data_dir <- file.path(system.file("extdata", package="daisyrVis"),
                      'annual/Annual-FN')
scenarios <- list('2-2b'='HourlyP-Annual-FN-2-2b.dlf',
                  '2-3b'='HourlyP-Annual-FN-2-3b.dlf',
                  '2-4b'='HourlyP-Annual-FN-2-4b.dlf',
                  '2-5b'='HourlyP-Annual-FN-2-5b.dlf')
dlfs <- lapply(scenarios, function(fname) {
    read_dlf(file.path(data_dir, fname))
})
# Be careful with variable names. R does not like '-' so we remove it
vars <- c('MatrixLeaching', 'CropUptake', 'SoilDrain', 'SurfaceLoss')
plot_annual(dlfs, vars, ' - Annual Field Nitrogen')
}
