% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlf.R
\name{read_dlf}
\alias{read_dlf}
\title{Read a dlf file (Daisy log file)}
\usage{
read_dlf(path)
}
\arguments{
\item{path}{Path to dlf file}
}
\value{
An S4 object of class Dlf with three slots: header, units, body

header is list with everything in the dlf file before the data. key: value
pairs from the header are stored as named components of the list.
units is a data.framecontaining the units of the values in body.
body is a data.frame containing the logged values
}
\description{
Read a dlf file (Daisy log file)
}
\examples{
data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, "annual/Annual-FN/HourlyP-Annual-FN-2-2b.dlf")
dlf <- read_dlf(path)
slotNames(dlf)
names(dlf@header)
dlf@units$Crop
dlf@body["Crop"]
}
